feat: gate DIGEST-MD5 and NTLM behind requirefips build tag
DIGEST-MD5 (MD5/MD4) and NTLM (MD4, HMAC-MD5, RC4, DES) use algorithms
that are not FIPS 140-2/3 approved and cannot be made compliant — the
protocols are inherently non-FIPS by design.

Under -tags requirefips, both mechanisms are replaced by stubs that
return a clear error directing users to GSSAPIBind (Kerberos) or
SimpleBind over TLS. This physically removes go-ntlmssp and the direct
x/crypto/md4 and crypto/md5 imports from the binary, satisfying the
strict FedRAMP High requirement that non-FIPS crypto must be absent
from the binary, not merely unreachable at runtime.

Non-FIPS builds are unchanged. A fips140.Enabled() runtime guard is
also added as a defence-in-depth safety net for binaries built without
the tag but running under GOFIPS140.

Known limitation: gokrb5 pulls in x/crypto/md4 and crypto/rc4 for its
RC4-HMAC Kerberos enctype support regardless of build tags. This will
be addressed in a separate upstream or Elastic fork PR.
